Refactor DirectHiddenStatesDataset and Trainer class
fea: add the layer idx select
This commit refactors the DirectHiddenStatesDataset and Trainer classes. It introduces a new parameter, "layer_idx", to the DirectHiddenStatesDataset class, allowing the user to specify which layer(s) to use. The "last_layer_only" parameter has been removed. In the Trainer class, the "layer_idx" argument is now passed to the DirectHiddenStatesDataset constructor. This change improves flexibility and code readability.